Time: 0 ms (100%), Space: 7.8 MB (65.87%) - LeetHub

+++ 3622-check-divisibility-by-digit-sum-and-product/3622-check-divisibility-by-digit-sum-and-product.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool checkDivisibility(int n) {
         int dibba = n;
-        int sum = 0;
+        long int sum = 0;
         int prod = 1;
         while(n>0){
             int dig = n%10;
